Move pop ups to shared components directory
