+++ src/components/admin-panel/venues-tab/venue-form/VenueForm.tsx
@@ -13,7 +13,7 @@ import { initializeVenueFormData } from "../../../../utils/utils";
 import TertiaryButton from "../../../shared-components/buttons/tertiary-button/TertiaryButton";
 import placeholderImage from "../../../../assets/upload-photo-placeholder.jpg";
 import axios from "axios";
-import InfoPopup from "../../new-movie/pop-up/InfoPopup";
+import InfoPopup from "../../../shared-components/pop-up/info-pop-up/InfoPopup";
 import LoadingIndicator from "../../../shared-components/loading-indicator/LoadingIndicator";
 import PrimaryButton from "../../../shared-components/buttons/primary-button/PrimaryButton";
 import SecondaryButton from "../../../shared-components/buttons/secondary-button/SecondaryButton";
